Test project - Basic test for clock and memory r/w - Peripherial test for SPI/I2C/UART

+++ test/inc/test_aes.h
@@ -51,7 +51,7 @@ Declaration of External Variables & Functions
 // Sec 4: declaration of external function prototype
 void TEST_AesInit(void);
 void TEST_AesSingleMode(void);
-void TEST_AesLoop(INT16 iCount);
+void TEST_AesLoop(int iCount);
 
 /******************************************************************************
 Declaration of data structure
